fix: launchdを実際に操作するテストをdarwin限定に分離
TestInstallLaunchAgent_WritesPlistAndBootstraps と
TestUninstallLaunchAgent_RemovesPlistAndBootsOut は launchctl 自体はスタブ
化しているが、launchAgentPlistPath が内部で使う os.UserHomeDir() は
Windows では $HOME ではなく %USERPROFILE% を見るため、
t.Setenv("HOME", tmpDir) によるリダイレクトが効かず実ユーザーの
プロファイル配下を触ってしまい、テストのアサーションも成立しなくなって
いた。launchd 自体がmacOS専有の仕組みでもあるため、この2つを
//go:build darwin の別ファイルに切り出した。スタブ型・ヘルパー・他の
launchd関連テスト（TestGenerateLaunchAgentPlist_ContainsExpectedFields、
TestUninstallLaunchAgent_MissingPlistIsNotAnError）は元のファイルのまま
変更していない。
