+++ server/app_agent_test.go
@@ -1,7 +1,6 @@
 package server
 
 import (
-	"os"
 	"path/filepath"
 	"strings"
 	"testing"
@@ -54,64 +53,6 @@ func TestGenerateLaunchAgentPlist_ContainsExpectedFields(t *testing.T) {
 		if !strings.Contains(content, want) {
 			t.Fatalf("expected generated plist to contain %q, got:\n%s", want, content)
 		}
-	}
-}
-
-func TestInstallLaunchAgent_WritesPlistAndBootstraps(t *testing.T) {
-	home := t.TempDir()
-	t.Setenv("HOME", home)
-
-	stub := &stubLaunchAgentRunner{}
-	withStubLaunchAgentRunner(t, stub)
-
-	if err := installLaunchAgent(); err != nil {
-		t.Fatalf("installLaunchAgent returned error: %v", err)
-	}
-
-	plistPath := filepath.Join(home, "Library", "LaunchAgents", "com.uxmusic.serve.plist")
-	data, err := os.ReadFile(plistPath)
-	if err != nil {
-		t.Fatalf("expected plist to be written to %s: %v", plistPath, err)
-	}
-	if !strings.Contains(string(data), "com.uxmusic.serve") {
-		t.Fatalf("written plist missing label, got:\n%s", string(data))
-	}
-
-	if len(stub.bootstrapCalls) != 1 {
-		t.Fatalf("expected exactly one Bootstrap call, got %d", len(stub.bootstrapCalls))
-	}
-	if stub.bootstrapCalls[0] != plistPath {
-		t.Fatalf("expected Bootstrap to be called with %s, got %s", plistPath, stub.bootstrapCalls[0])
-	}
-	if len(stub.bootoutCalls) != 0 {
-		t.Fatalf("installLaunchAgent must not call Bootout, got %v", stub.bootoutCalls)
-	}
-}
-
-func TestUninstallLaunchAgent_RemovesPlistAndBootsOut(t *testing.T) {
-	home := t.TempDir()
-	t.Setenv("HOME", home)
-
-	stub := &stubLaunchAgentRunner{}
-	withStubLaunchAgentRunner(t, stub)
-
-	if err := installLaunchAgent(); err != nil {
-		t.Fatalf("installLaunchAgent returned error: %v", err)
-	}
-	plistPath := filepath.Join(home, "Library", "LaunchAgents", "com.uxmusic.serve.plist")
-	if _, err := os.Stat(plistPath); err != nil {
-		t.Fatalf("expected plist to exist before uninstall: %v", err)
-	}
-
-	if err := uninstallLaunchAgent(); err != nil {
-		t.Fatalf("uninstallLaunchAgent returned error: %v", err)
-	}
-
-	if _, err := os.Stat(plistPath); !os.IsNotExist(err) {
-		t.Fatalf("expected plist to be removed after uninstall, stat err: %v", err)
-	}
-	if len(stub.bootoutCalls) != 1 || stub.bootoutCalls[0] != launchAgentLabel {
-		t.Fatalf("expected exactly one Bootout call with label %s, got %v", launchAgentLabel, stub.bootoutCalls)
 	}
 }
